Add user if user currently does not exist.
