--- src/test/java/com/google/sps/servlets/CreateUser.java
@@ -1,58 +1,0 @@
-package com.google.sps.servlets;
-
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
-import com.google.appengine.api.datastore.Entity;
-import com.google.appengine.api.datastore.EntityNotFoundException;
-import com.google.appengine.api.datastore.Key;
-import com.google.appengine.api.datastore.KeyFactory;
-import com.google.appengine.api.datastore.Transaction;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseAuthException;
-import com.google.firebase.auth.UserRecord;
-import com.google.sps.firebase.FirebaseAppManager;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.ConcurrentModificationException;
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-// Once class owner submits a TA email, retrieve that user and add them as a TA to the class
-@WebServlet("/create-user")
-public class CreateUser extends HttpServlet {
-
-  private FirebaseAuth authInstance;
-
-  // Get the current session
-  @Override
-  public void init(ServletConfig config) throws ServletException {
-    try {
-      authInstance = FirebaseAuth.getInstance(FirebaseAppManager.getApp());
-    } catch (IOException e) {
-      throw new ServletException(e);
-    }
-  }
-
-  // Add a user to the datastore
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    
-    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-
-    // Check if user entity already exists
-    String userToken = request.getParameter("userToken");
-    FirebaseToken decodedToken = authInstance.verifyIdToken(userToken);
-    String userID = decodedToken.getUid();
-
-    UserRecord userRecord = authInstance.getUser(userID);
-    String userEmail = userRecord.getEmail();
-
-    
-
-    // Create a new user entity if there is no existing one
-  }
-}
